+++ src/libc/freestanding/system_calls/waitForServiceRegistered.cpp
@@ -11,7 +11,7 @@ void waitForServiceRegistered(Kernel::ServiceType type) {
         IPC::MaximumMessageBuffer buffer;
         receive(&buffer);
 
-        if (buffer.messageNamespace == static_cast<uint32_t>(IPC::MessageNamespace::ServiceRegistry)
+        if (buffer.messageNamespace == IPC::MessageNamespace::ServiceRegistry
                 && buffer.messageId == static_cast<uint32_t>(Kernel::MessageId::NotifyServiceRegistered)) {
             auto notify = IPC::extractMessage<Kernel::NotifyServiceRegistered>(buffer);
